[add] functionality for `/cupcakes/avg/rating/<cupcake_id>`
Utilised the `aggregate` functionality in mongoDB to  match and group all ratings documents in the ratings collection based on the given `cupcakeId`. Further improvements would be to round the value of `avgRating` to 2 decimal places, and more importantly, not use the DB aggregate function.

By utilising the DB aggregate function, I'm adding additional complexity to the query and comsuming more RAM and CPU. This query should actually grab a list of all appropriate ratings and the calculcation be completed on the backend application. Backend applications are easier to scale than databases and if this was a production application it will increase latency as traffic spikes.